+++ app/src/main/java/com/ttt/qx/qxcall/function/login/model/entity/UserListInfo.java
@@ -153,7 +153,7 @@ public class UserListInfo implements Serializable{
             private String nick_name;
             private String member_avatar;
             private int level;
-            private int member_account;
+            private double member_account;
             private String member_sex;
             private String member_age;
             private int member_price;
@@ -204,11 +204,11 @@ public class UserListInfo implements Serializable{
                 this.level = level;
             }
 
-            public int getMember_account() {
+            public double getMember_account() {
                 return member_account;
             }
 
-            public void setMember_account(int member_account) {
+            public void setMember_account(double member_account) {
                 this.member_account = member_account;
             }
